- dynamic computation graph (beta) 	- add supports (/express) 	- add tests 	- add benchmarks with it (/benchmark/exprModels) - Python 	- MNN engine and tools were submitted to pip 	- available on Windows/macOS/Linux - Engine/Converter 	- add supports for each op benchmarking 	- refactor optimizer by separating steps - CPU 	- add supports for Conv3D, Pool3D, ELU, ReverseSequence 	- fix ArgMax, Permute, Scale, BinaryOp, Slice, SliceTf - OpenCL 	- add half transform in CPU 	- add broadcast supports for binary 	- optimize Conv2D, Reshape, Eltwise, Gemm, etc. - OpenGL 	- add sub, real div supports for binary 	- add supports for unary 	- optimize Conv2D, Reshape - Vulkan 	- add max supports for eltwise - Metal 	- fix metallib missing problem - Train/Quantization 	- use express to refactor training codes

+++ source/shape/ShapeExpandDims.cpp
@@ -14,7 +14,7 @@ class ExpandDimsComputer : public SizeComputer {
 public:
     virtual bool onComputeSize(const MNN::Op* op, const std::vector<Tensor*>& inputs,
                                const std::vector<Tensor*>& outputs) const override {
-        const int inputSize = inputs.size();
+        const int inputSize = (int)inputs.size();
         MNN_ASSERT(2 == inputSize || 1 == inputSize);
         MNN_ASSERT(1 == outputs.size());
